[FIX JENKINS-51667] Move getBuildFingerprints from AbstractBuild to Run
